requirements.txt
amqp==5.1.1
async-timeout==4.0.2
base58==2.1.1
billiard==3.6.4.0
celery==5.2.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.0
ecdsa==0.15
Flask==2.2.2
gunicorn==20.1.0
h11==0.13.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4
pymysql==1.0.2
MarkupSafe==2.1.1
packaging==21.3
parsimonious==0.8.1
prometheus-client==0.16.0
prompt-toolkit==3.0.30
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
toolz==0.12.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
anyio==4.0.0
base58==2.1.1
certifi==2023.7.22
cffi==1.16.0
cryptography==41.0.5
Deprecated==1.2.14
ECPy==1.2.5
exceptiongroup==1.1.3
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.21
pycryptodome==3.19.0
sniffio==1.3.0
types-Deprecated==1.2.9.3
typing_extensions==4.8.0
websockets==10.4
wrapt==1.16.0
xrpl-py==2.4.0
